+++ include/Makefile.am
@@ -1,2 +1,2 @@
 include_HEADERS = xpmem.h
-
+noinst_HEADERS = xpmem_internal.h
