+++ kernel/xpmem_attach.c
@@ -17,7 +17,7 @@
 #include <linux/mm.h>
 #include <linux/mman.h>
 #include <linux/file.h>
-#include <xpmem.h>
+#include "xpmem_internal.h"
 #include "xpmem_private.h"
 
 static void
